deleted unused files, separated code into different files
