+++ .vscode/c_cpp_properties.json
@@ -4,6 +4,7 @@
             "name": "Win32",
             "includePath": [
                 "${workspaceFolder}/**",
+                "${workspaceFolder/inc/**",
                 "${workspaceFolder}/libopencm3/include/**"
             ],
             "defines": [
