--- my-common-code/api-asm.h
@@ -1,7 +1,0 @@
-#include <stdint.h>
-/**
- * Reverse bytes in words
- * @param in word in form abcd
- * @returns word in form dcba
- */
-uint32_t rev_bytes(uint32_t in);
