--- my-common-code/api.c
@@ -1,7 +1,0 @@
-
-#include "api.h"
-
-int my_func(int a)
-{
-	return a * 3;
-}
