--- my-common-code/api.h
@@ -1,1 +1,0 @@
-int my_func(int a);
